requirements.txt
numpy
pandas
opencv-python
mediapipe
tensorflow
keras
scikit-learn
tk
Pillow
